chore(Kissmetrics): move sdk loader to different file
